Fix player rotation stuttering
Instead of relying on the player's Rigidbody for interpolation, create
separate AstronautVisual and AstronautSimulated objects for the player
and handle transform interpolation in scripts.

Unity's Rigidbody has a useful `interpolation` property. This helps
prevent physics stuttering that would otherwise be apparent due to
positions and rotations only being updated in FixedUpdate instead of
every render.

However, this interpolation doesn't work if you are using
`MovePosition()` or `MoveRotation()` with a non-kinematic Rigidbody,
since the interpolation is meant for simulated physics. Kinematic
rigidbodies do get interpolation from these functions, but that disables
physics simulation.

This caused problems for our player character, which updates its
position based on physics forces but updates its rotation using
`MoveRotation()`. We want our character to get pushed and pulled by
physics forces, but we want to directly set the rotation (and
interpolate to make it look nice). We also don't want to change game
state in Update() to keep gameplay independent from framerate.

The solution: create separate objects for the visual and simulated
parts of the player, and add an InterpolatedTransform script to the
visual part to handle interpolation outside of the Rigidbody.

This solution was based on this Reddit thread:
https://www.reddit.com/r/Unity3D/comments/9kofzk/looking_for_some_help_with_interpolating/e72vhaw/?context=10000
Interpolation code was taken from this extremely helpful article:
https://www.kinematicsoup.com/news/2016/8/9/rrypp5tkubynjwxhxjzd42s3o034o8