New player model (#22)
* WIP: Big animation refactor

* Various animation and movement fixes

- Don't apply gravity twice to player
- Increase grounded down force
- Update animation controller

* PlayerMovement: Cleanup and refactor

Move getting the move direction into a separate function.

* Rename animation controller, remove exit times

* Refactor: remove AnimatorManager, cleanup InputManager

* Add weapon model and rig with IK

The player now holds a raygun. This is currently at a fixed position
relative to the player until we add scripts or animation.

* Adjust gun placement

* Make Raygun child of right hand

Now the ref_grip tells us where to aim and we don't get weird issues
with the raygun not moving with the hand.

* Add Stylized Astronaut asset and test scene

This one seems easier to work with than the existing LowPolyMan.

* Add head look rig

* Add existing player scripts to Stylized Astronaut

Animations are broken until I make a new controller.

* Revert Astronaut to generic

This model is incompatible with the Humanoid rig.

* Add new animation controller

We don't have a death animation, and this isn't a humanoid avatar! Darn.
We can make one in the future, but for now I'm using the Flip animation.

* Fix collider placement

Fixes player floating.

* Add gun and aim at target

The arm animations get a bit wonky when you're aiming to the astronaut's
left, but it's mostly reasonable. I also moved the projectile spawn
point to where the gun is.

* Make new astronaut into prefab

* Add new astronaut to all scenes

* Add laser VFX to Astronaut prefab