Add Gravity Plane With Trigger
This is a gravity plane that is only active while the player is in the
trigger zone. This change also lets us enable/disable gravity sources
via scripts.

This will cause other CustomGravityRigidbody objects in the trigger zone
to not be affected by gravity when the player leaves, which is weird.
Ideally we'd have a separate list of sources for each rigidbody.