Make jetpack apply continuous force
Dashing and air jumping are replaced with a jetpack that applies a
continuous force. For dashing, the force is in your movement direction,
and air jumping now applies force upwards. This change is intended to
make it easier to make precise movements in a way that fits with the
gravity system.